requirements.txt
coverage
dateparser
flasgger
Flask
Flask-Cors
PyYAML==5.3.1
PyJWT==1.7.1
pymongo
pytest
python-dateutil
pytz
requests
sendgrid
xlrd
oauth2client==2.2.0
bcrypt
pycrypto
flask-restx